requirements.txt
# Core Dependencies
discord.py==2.4.0
azure-cosmos==4.7.0
python-dotenv==1.0.0
pydantic==2.5.2

# Async Support
aiohttp==3.9.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Development Tools
black==24.10.0
flake8==7.0.0
isort==5.13.2
pre-commit==3.6.0
pip-tools==7.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
mypy==1.7.1

# Type Checking
typing-extensions>=4.9.0
types-python-jose

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Security
cryptography==41.0.5
python-jose==3.3.0
python-dateutil==2.8.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
